backend/requirements.txt
accelerate==1.11.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
bitsandbytes==0.48.2
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
distro==1.9.0
faiss-cpu==1.12.0
fastapi==0.120.4
filelock==3.20.0
fsspec==2025.10.0
groq==0.33.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.5
numpy==2.3.4
packaging==25.0
pillow==12.0.0
psutil==7.1.3
pydantic==2.12.3
pydantic_core==2.41.4
pypdf==6.1.3
python-dotenv==1.2.1
python-multipart==0.0.20
PyYAML==6.0.3
regex==2025.10.23
requests==2.32.5
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.3
sentence-transformers==5.1.2
sniffio==1.3.1
starlette==0.49.3
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.9.0+cu126
torchvision==0.24.0+cu126
tqdm==4.67.1
transformers==4.57.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
